+++ CHANGELOG.md
@@ -10,12 +10,13 @@ DEPS_CLJ_TOOLS_VERSION=1.11.1.1165 bb clojure
 
 [deps.clj](https://github.com/borkdude/deps.clj): a faithful port of the clojure CLI bash script to Clojure
 
-## 1.11.1.1386
+## Unreleased
 
 - [#104](https://github.com/borkdude/deps.clj/issues/104): print repl-aliases in -Sdescribe
 - Make installation of tools jar more robust using transaction file
 - [#106](https://github.com/borkdude/deps.clj/issues/106): System deps.edn should be extracted on tools install
 - Download tools jar from new Github releases link
+- Verify downloaded zip file with .sha256 file
 
 ## 1.11.1.1347